pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pdf-ocr-md"
version = "0.1.0"
description = "OCR PDFs with Ollama and generate clean markdown outputs"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "httpx>=0.27.0",
  "pypdf>=5.0.0",
  "pypdfium2>=4.30.0",
  "Pillow>=10.0.0",
  "rich>=13.0.0",
]

[project.scripts]
pdf-ocr-md = "pdf_ocr_md.main:main"
pdf-ocr-md-tui = "pdf_ocr_md.interactive_tui:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
